+++ Program/Validators/end_of_signature_space_validator.py
@@ -38,8 +38,7 @@ class EndOfSignatureSpaceValidator:
                     error_dictionary["error_list"] \
                         .append("End Of Signature Error: A space is needed in"
                                 + f" '{bracket_end_entry.strip()}'" 
-                                + " between the ) and {" 
-                                + f" on line {line_number}.")
+                                + " between the ) and {.")
                     error_dictionary["error_count"] += 1
 
             # Then check signatures with an else
