+++ Program/Validators/start_of_signature_spaceValidator.py
